passa-a-bola-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.2.0
python-multipart==0.0.17
python-dotenv==1.0.1

